requirements.txt
dnspython
tabulate
requests